+++ src/components/astral-grid/astral-grid-item/test/astral-grid-item.spec.tsx
@@ -1,0 +1,18 @@
+import { newSpecPage } from '@stencil/core/testing';
+import { AstralGridItem } from '../astral-grid-item';
+
+describe('astral-grid-item', () => {
+  it('renders', async () => {
+    const page = await newSpecPage({
+      components: [AstralGridItem],
+      html: `<astral-grid-item></astral-grid-item>`,
+    });
+    expect(page.root).toEqualHtml(`
+      <astral-grid-item>
+        <mock:shadow-root>
+          <slot></slot>
+        </mock:shadow-root>
+      </astral-grid-item>
+    `);
+  });
+});