+++ CHANGELOG.md
@@ -4,7 +4,15 @@
 
 Following Material Design 3 specs
 
+### Version 0.0.8
+
+- Added astral-grid
+- Added astral-layout
+- Added astral-container
+- Added first documentation page (buttons)
+
 ### Version 0.0.7
+
 - Gzipped dist files
 - Added icon button
 - Added components
@@ -12,28 +20,34 @@ Following Material Design 3 specs
   - Modals
 
 ### Version 0.0.6
+
 - Prevented icon selection from astral-checkboxes
 - Patched README & documentation
 - Patched SCSS files not present in the published package
 
 ### Version 0.0.5
+
 - Patched wrong loader path
 - Patched missing SCSS ressources in dist folder
 
 ### Version 0.0.4
+
 - Patched CSS variables into a reusable SCSS mixin that transforms your own palette (in HEX) into an rgb map, that can be used to opacify colors of your theme more easily
 
 ### Version 0.0.3
+
 - Added components
   - Breadcrumb
   - Checkboxes
   - Dividers
 
 ### Version 0.0.2
+
 - Added components
   - Accordion
 
-### Version 0.0.1 
+### Version 0.0.1
+
 - Initial release
 - Added components
   - Button